Extract Analysis Subheader to its own component.
